socialmedia-backend/requirements.txt
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0